fix incorrect type conversions via static_cast

+++ src/json/SimdJsonApixu.cpp
@@ -81,7 +81,7 @@ bool SimdJsonApixu::parseCurrentWeather(const std::string& json, Weather& weathe
     current["pressure_mb"].tie(v2, error);
     if (!error && v2.is<double>())
     {
-      weather.setPressure(static_cast<float>(v2.get<double>().value()));
+      weather.setPressure(static_cast<int16_t>(v2.get<double>().value()));
     }
     // cloudiness
     current["cloud"].tie(v2, error);
